🏗️ (dotori-note): refactor project structure
