+++ apps/dotori-note/notes/ui-craft/list-item-expand-transition/list-item-expand-transition.mdx
@@ -6,8 +6,8 @@ created: 2025-04-01
 ---
 
 import { ListItemExpandTransition } from '@siluat/ui-craft/react/list-item-expand-transition';
-import UiCraftFrame from '@/entities/note/ui/UiCraftFrame.astro';
-import SquareEnixCopyrightAlert from '@/entities/note/ui/SquareEnixCopyrightAlert.astro';
+import UiCraftFrame from '@/note-ui/UiCraftFrame.astro';
+import SquareEnixCopyrightAlert from '@/note-ui/SquareEnixCopyrightAlert.astro';
 
 목록의 아이템을 클릭/터치하면 해당 아이템의 상세 정보를 확장 트랜지션으로 보여준다.
