+++ apps/dotori-note/notes/ui-craft/svg-path-morphing/svg-path-morphing.mdx
@@ -8,8 +8,8 @@ description: 코드로 SVG Morphing 제어하기
 
 import { FfxivJobIconsLoopUsingGooeyEffect } from '@siluat/ui-craft/react/ffxiv-job-icons-loop-using-gooey-effect';
 import { FfxivJobIconsLoopUsingFlubber } from '@siluat/ui-craft/react/ffxiv-job-icons-loop-using-flubber';
-import UiCraftFrame from '@/entities/note/ui/UiCraftFrame.astro';
-import SquareEnixCopyrightAlert from '@/entities/note/ui/SquareEnixCopyrightAlert.astro';
+import UiCraftFrame from '@/note-ui/UiCraftFrame.astro';
+import SquareEnixCopyrightAlert from '@/note-ui/SquareEnixCopyrightAlert.astro';
 
 ## 개요
