+++ apps/dotori-note/src/meta/Favicon.astro
