+++ apps/dotori-note/src/note-ui/SquareEnixCopyrightAlert.astro
