+++ apps/dotori-note/src/note-ui/UiCraftFrame.astro
