+++ apps/dotori-note/src/pages/index.astro
@@ -1,10 +1,10 @@
 ---
-import DefaultLayout from '@/page-layouts/DefaultLayout.astro';
-import LastModifiedNotes from '@/features/last-modified-notes/ui/LastModifiedNotes.astro';
+import DefaultPageLayout from '@/layouts/DefaultPageLayout.astro';
+import LastModifiedNotes from '@/note-ui/LastModifiedNotes.astro';
 import { RandomQuote } from '@/widgets/random-quote/ui/RandomQuote';
 ---
 
-<DefaultLayout title="도토리 노트">
+<DefaultPageLayout title="도토리 노트">
   <section class="m-4">
     <h1 class="text-xs mb-1 font-bold text-zinc-500 dark:text-zinc-400">최근 변경된 노트</h1>
     <LastModifiedNotes />
@@ -12,4 +12,4 @@ import { RandomQuote } from '@/widgets/random-quote/ui/RandomQuote';
   <section class="m-4">
     <RandomQuote client:only="react" />
   </section>
-</DefaultLayout>
+</DefaultPageLayout>
