+++ apps/dotori-note/src/pages/note/[id].astro
@@ -1,7 +1,7 @@
 ---
-import ContentLayout from '@/page-layouts/ContentLayout.astro';
+import ContentPageLayout from '@/layouts/ContentPageLayout.astro';
 import { getCollection, render } from 'astro:content';
-import Blockquote from '@/mdx-custom-ui/Blockquote.astro';
+import Blockquote from '@/note-ui/Blockquote.astro';
 
 export async function getStaticPaths() {
   const notes = await getCollection('notes');
@@ -15,11 +15,11 @@ const { note } = Astro.props;
 const { Content } = await render(note);
 ---
 
-<ContentLayout 
+<ContentPageLayout 
   title={note.data.title} 
   description={note.data.description} 
   created={note.data.created} 
   modified={note.data.modified}
 >
   <Content components={{ blockquote: Blockquote }} />
-</ContentLayout>
+</ContentPageLayout>
