+++ apps/dotori-note/src/shared-types/utility-types.ts
