+++ apps/dotori-note/src/styles/InitTheme.astro
