+++ apps/dotori-note/src/styles/global.css
