+++ apps/dotori-note/src/widgets/random-quote/lib/pick-random-quote.ts
@@ -1,4 +1,4 @@
-import type { NonEmptyArray } from '../../../shared/model/utility-types';
+import type { NonEmptyArray } from '@/shared-types/utility-types';
 import type { Quote } from '../model/quote-types';
 
 export function pickRandomQuote(quotes: NonEmptyArray<Quote>): Quote {
